requirements.txt
selenium==4.15.2
pyYAML==6.0.1